requirements.txt
Flask==2.0.1
folium==0.12.1.post1
oauth2client==4.1.3
pandas
protobuf==4.22.1
python-dotenv==1.0.0
requests==2.25.0
google-api-core==2.11.0
google-api-python-client==2.63.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.6.0
google-cloud-bigquery-storage==2.18.1
google-cloud-core==2.3.2
google-cloud-logging==3.5.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.6
pydata-google-auth==1.7.0